refactor: extract createFilterHelpers util addressing quality gate CI
