+++ src/components/BenefitPlanBeneficiariesFilter.jsx
@@ -1,13 +1,13 @@
 import React from 'react';
-import _debounce from 'lodash/debounce';
 import { injectIntl } from 'react-intl';
 import { Grid } from '@mui/material';
 import { styled } from '@mui/material/styles';
 import {
   TextInput, PublishedComponent, formatMessage, ConstantBasedPicker, GRID_RESPONSIVE_STANDARD,
 } from '@openimis/fe-core';
-import { CONTAINS_LOOKUP, DEFAULT_DEBOUNCE_TIME, EMPTY_STRING } from '../constants';
+import { CONTAINS_LOOKUP } from '../constants';
 import { defaultFilterStyles } from '../util/styles';
+import { createFilterHelpers } from '../util/filter-helpers';
 import BeneficiaryStatusPicker from '../pickers/BeneficiaryStatusPicker';
 
 const StyledGrid = styled(Grid)(({ theme }) => ({
@@ -23,31 +23,7 @@ function BenefitPlanBeneficiariesFilter({
 }) {
   const any = formatMessage(intl, 'socialProtection', 'any');
 
-  const debouncedOnChangeFilters = _debounce(onChangeFilters, DEFAULT_DEBOUNCE_TIME);
-
-  const filterValue = (filterName) => filters?.[filterName]?.value;
-
-  const filterTextFieldValue = (filterName) => filters?.[filterName]?.value ?? EMPTY_STRING;
-
-  const onChangeStringFilter = (filterName, lookup = null) => (value) => {
-    if (lookup) {
-      debouncedOnChangeFilters([
-        {
-          id: filterName,
-          value,
-          filter: `${filterName}_${lookup}: "${value}"`,
-        },
-      ]);
-    } else {
-      onChangeFilters([
-        {
-          id: filterName,
-          value,
-          filter: `${filterName}: "${value}"`,
-        },
-      ]);
-    }
-  };
+  const { filterValue, filterTextFieldValue, onChangeStringFilter } = createFilterHelpers(filters, onChangeFilters);
 
   return (
     <StyledGrid container>
